+++ Network/layers/OutputLayer.h
@@ -12,7 +12,7 @@ namespace cn {
     public:
         OutputLayer(int id, cn::Network &network);
         void run(const Bitmap<float> &input) override;
-        float getChain(const Vector3<float> &input) override;
+        float getChain(const Vector3<int> &input) override;
 
     };
 }
